requirements.txt
python-chess==1.999
fastapi==0.115.0
uvicorn[standard]==0.32.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2